# server/docs/swagger.yaml
basePath: /
definitions:
...
  model.AvfCard:
    properties:
      author:
        description: 作者
        type: string
      contractAddress:
        description: 合约地址
        type: string
      createDate:
        description: 创建时间
        type: string
      createdAt:
        description: 创建时间
        type: string
      desc:
        description: 描述
        type: string
      id:
        description: 主键ID
        type: integer
      level:
        description: 等级 1-N 2-R 3-SR 4-SSR
        type: integer
      money:
        description: 价格
        type: number
      name:
        description: 卡牌名称
        type: string
      number:
        description: 发行数量
        type: integer
      star:
        description: 算力值
        type: integer
      status:
        description: 状态 true-正常 false-禁用
        type: boolean
      updateDate:
        description: 修改时间
        type: string
      updatedAt:
        description: 更新时间
        type: string
      walletAddress:
        description: 收款钱包地址
        type: string
    type: object
...
  model.AvfOrder:
    properties:
      block:
        description: 区块号
        type: string
      createdAt:
        description: 创建时间
        type: string
      from:
        description: 支付地址
        type: string
      gas:
        description: 手续费
        type: string
      gasPrice:
        description: 手续费价格
        type: string
      id:
        description: 主键ID
        type: integer
      num:
        description: 数量
        type: integer
      number:
        description: 剩余数量
        type: integer
      orderSn:
        description: 订单编号
        type: string
      payTime:
        description: 支付时间
        type: integer
      price:
        description: 价格
        type: integer
      status:
        description: 状态 1-待支付 2-待确认 3-已完成 4-已取消5-已关闭
        type: integer
      to:
        description: 收款地址
        type: string
      txHash:
        description: 交易hash
        type: string
      uid:
        description: 用户ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  model.AvfOrderCard:
    properties:
      cardId:
        description: 卡牌ID
        type: integer
      createTime:
        description: 创建时间
        type: integer
      createdAt:
        description: 创建时间
        type: string
      giveType:
        description: 获得方式 1-抽奖 2-购买
        type: integer
      id:
        description: 主键ID
        type: integer
      level:
        description: 等级 1-N 2-R 3-SR 4-SSR
        type: integer
      orderId:
        description: 订单ID
        type: integer
      star:
        description: 算力值
        type: integer
      status:
        description: 状态 1-正常 2-转让中
        type: integer
      uid:
        description: 用户ID
        type: integer
      updateTime:
        description: 修改时间
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  request.AvfOrderCardSearch:
    properties:
      cardId:
        description: 卡牌ID
        type: integer
      createTime:
        description: 创建时间
        type: integer
      createdAt:
        description: 创建时间
        type: string
      giveType:
        description: 获得方式 1-抽奖 2-购买
        type: integer
      id:
        description: 主键ID
        type: integer
      level:
        description: 等级 1-N 2-R 3-SR 4-SSR
        type: integer
      orderId:
        description: 订单ID
        type: integer
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      star:
        description: 算力值
        type: integer
      status:
        description: 状态 1-正常 2-转让中
        type: integer
      uid:
        description: 用户ID
        type: integer
      updateTime:
        description: 修改时间
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
  request.AvfOrderSearch:
    properties:
      block:
        description: 区块号
        type: string
      createdAt:
        description: 创建时间
        type: string
      from:
        description: 支付地址
        type: string
      gas:
        description: 手续费
        type: string
      gasPrice:
        description: 手续费价格
        type: string
      id:
        description: 主键ID
        type: integer
      num:
        description: 数量
        type: integer
      number:
        description: 剩余数量
        type: integer
      orderSn:
        description: 订单编号
        type: string
      page:
        description: 页码
        type: integer
      pageSize:
        description: 每页大小
        type: integer
      payTime:
        description: 支付时间
        type: integer
      price:
        description: 价格
        type: integer
      status:
        description: 状态 1-待支付 2-待确认 3-已完成 4-已取消5-已关闭
        type: integer
      to:
        description: 收款地址
        type: string
      txHash:
        description: 交易hash
        type: string
      uid:
        description: 用户ID
        type: integer
      updatedAt:
        description: 更新时间
        type: string
    type: object
...
  request.IdsReq:
    properties:
      ids:
        items:
          type: integer
        type: array
    type: object
...
  web_tools.CardListResponse:
    properties:
      list:
        description: 卡牌列表
        items:
          $ref: '#/definitions/model.AvfCard'
        type: array
      total:
        description: 总条数
        type: integer
    type: object
...
info:
  contact: {}
  description: This is a sample Server pets
  title: Swagger Example API
  version: 0.0.1
paths:
...
  /avfOrder/createAvfOrder:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建AvfOrder
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AvfOrder'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建AvfOrder
      tags:
      - AvfOrder
  /avfOrder/deleteAvfOrder:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除AvfOrder
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AvfOrder'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除AvfOrder
      tags:
      - AvfOrder
  /avfOrder/deleteAvfOrderByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除AvfOrder
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除AvfOrder
      tags:
      - AvfOrder
  /avfOrder/findAvfOrder:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询AvfOrder
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AvfOrder'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询AvfOrder
      tags:
      - AvfOrder
  /avfOrder/getAvfOrderList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页获取AvfOrder列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AvfOrderSearch'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取AvfOrder列表
      tags:
      - AvfOrder
  /avfOrder/updateAvfOrder:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新AvfOrder
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AvfOrder'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新AvfOrder
      tags:
      - AvfOrder
  /avfOrderCard/createAvfOrderCard:
    post:
      consumes:
      - application/json
      parameters:
      - description: 创建AvfOrderCard
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AvfOrderCard'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 创建AvfOrderCard
      tags:
      - AvfOrderCard
  /avfOrderCard/deleteAvfOrderCard:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 删除AvfOrderCard
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AvfOrderCard'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 删除AvfOrderCard
      tags:
      - AvfOrderCard
  /avfOrderCard/deleteAvfOrderCardByIds:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 批量删除AvfOrderCard
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.IdsReq'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"批量删除成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 批量删除AvfOrderCard
      tags:
      - AvfOrderCard
  /avfOrderCard/findAvfOrderCard:
    get:
      consumes:
      - application/json
      parameters:
      - description: 用id查询AvfOrderCard
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AvfOrderCard'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"查询成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 用id查询AvfOrderCard
      tags:
      - AvfOrderCard
  /avfOrderCard/getAvfOrderCardList:
    get:
      consumes:
      - application/json
      parameters:
      - description: 分页获取AvfOrderCard列表
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/request.AvfOrderCardSearch'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"获取成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 分页获取AvfOrderCard列表
      tags:
      - AvfOrderCard
  /avfOrderCard/updateAvfOrderCard:
    put:
      consumes:
      - application/json
      parameters:
      - description: 更新AvfOrderCard
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/model.AvfOrderCard'
      produces:
      - application/json
      responses:
        "200":
          description: '{"success":true,"data":{},"msg":"更新成功"}'
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: 更新AvfOrderCard
      tags:
      - AvfOrderCard
...
  /web/card/list:
    get:
      consumes:
      - application/json
      parameters:
      - description: 页码
        in: query
        name: page
        type: string
      - description: 数量
        in: query
        name: size
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/web_tools.CardListResponse'
      summary: 获取卡牌列表
      tags:
      - 前端接口
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: x-token
    type: apiKey
swagger: "2.0"